리소스 UI및 기능 구현

+++ Source/JH_Cropout/Public/Save/JHSaveGame.h
@@ -22,7 +22,7 @@ public:
 	FRandomStream Seed;
 
 	UPROPERTY()
-	TMap<EResourceType,int64> Resources;
+	TMap<EResourceType,int32> Resources;
 
 	UPROPERTY()
 	TArray<FSaveInteract> Interactables;
